BUG: rollback using executescripts in create_spatialdb (#628)
Recently the windows CI tests started crashing regularly. The crashes
seem to be happening in `sqlite_util.create_spatialdb`, but the crash
happens on different moments in different tests, so it is not
consistent.

The only recent change in the function is to use `executescript` instead
of `execute`, not sure how this would lead to occasional crashes in this
function, but his PR rolls back this change hoping it makes the
function/tests stable again.