ENH: in add_column, don't add column if update expression is invalid (#650)
If adding the column succeeds but the update fails, the output files is
in a state that is not ideal: when the `add_column` is ran again, the
column won't be "fixed". It would be better that this is transactional:
if the update fails, the add should be rolled back well.

This PR adds using transactions for this case.

Mind: transactions are not supported for all file types! E.g. 
- for .gpkg this will work fine.
- for .shp the column will still stay added as transactions are not
supported.