ENH: use forkserver to create worker processes on linux (#675)
Use "forkserver" to create worker processes on linux.

This solves the following deprecationwarning: `This process (pid=2361)
is multi-threaded, use of fork() may lead to deadlocks in the child.`

"forkserver" will become the default on [python
3.14](https://docs.python.org/3.14/whatsnew/3.14.html#concurrent-futures).

Note: this does have a significant performance impact in some use cases.
E.g. running the tests (a lot of operations on small files that are
forced to run in 2 processes) took:
- 1.5 minutes using fork
- 5.5 minutes using forkserver
- 7.5 minutes using spawn
- 1.5 minutes using threads (which is now the case: if input files < 100
rows)