BUG: on some input files spatial operations take significant time even though output exists (#696)
The idea is that if the output file exists already and `force=False`,
that no processing is done and hence dealing with that is fast.

However, if the input geopackages don't have a filled out
gpkg_ogr_contents.featurecount table/column, in many cases this seems to
take quite some time anyway.

This PR moves the output_exists check forward on many places to speed
this up. Disadvantage is that most input parameter checks are now
afterwards...

reference #695