ENH: add include_duplicates parameter to overlay operations (#757)
With the `include_duplicates` parameter added to `union`, `intersection`
and `identity`, it is possible to choose whether the output contains
both combinations column value directions (leading to duplicate
geometries in the output) or only one.

The default is `include_duplicates=True`, which keeps the default
behaviour the same as before.