ENH: for operations with steps, put all temp files in single temp dir (#761)
For operation with several steps (e.g. `union`,
`dissolve_within_distance`,...) now there are several temp dirs created
in the "root" geofileops temp dir, which is not super handy to debug.

This PR changes this so a temp directory is created for the main
operation, and all steps will create subdirectories in that directory to
store temp their specific temp files.

Also some related cleanup:
- use a context handler to create and cleanup the temp directories
instead of try/finally
- improve test coverage
- ...