ENH: add add_columns function (#768)
Function to add multiple columns in one "transaction", including giving
them a value if wanted.

In practice, a temp, local copy is taken from the input file, and the
result is copied to the output_path. Because the adding is done on a
temporary copy, some hacks can be used to speed it up, like disabling
the sqlite journalling.
This gives a very significant performance improvement.