ENH: add configuration option to remove slivers (#776)
For all spatial overlays, sliver polygons are removed from the output by
default.

Polygons are considered slivers if they are narrower than a certain
tolerance. By default this tolerance is 0.001 (CRS units) if the CRS of
the input layers is a projected CRS, 1e-7 if it is a geographic CRS.

You can change this default tolerance using the
`options.set_sliver_tolerance` runtime option.

resolves #774