ENH: if a function fails because the file is locked, retry a few times (#785)
At least when using networked storage, file locks can take some time to
be removed.

This causes some functions to fail occasionally (~1 in 500 calls based
on a specific test) when ran on a file located on network storage.

This PR implements a retry to avoid errors due to this phenomenon for
all functions that update an existing file (except for `copy_layer` for
now):
- `execute_sql`
- `create_spatial_index`
- `remove_spatial_index`
- `rename_layer`
- `rename_column`
- `add_column`
- `add_columns`
- `drop_column`
- `update_column`

To be evaluated if it should also be applied to read-only functions.